Repository state used for vCHEP 2021 paper results: * [Feature] Interactive track viewer improved * [Refactored] r-z / x-y map generation * [Refactored] weighted graph navigator now up-to-date * [Feature] script which generates plot for paper added

+++ utilities/interactive_track_viewer/TrackListWidgetItem.hpp
@@ -11,7 +11,7 @@
 
 struct TrackListWidgetItem : public QListWidgetItem
 {
-    using TrackPtr = const std::pair<float, std::span<RowTuple>> *;
+    using TrackPtr = const std::pair<double, std::span<RowTuple>> *;
 
     const TrackPtr track = nullptr;
